refactor container builder to pkg/common
